2.리팩토링
- 하위 컴포넌트 별도 파일로 분리
